+++ app/api/views/__init__.py
@@ -1,0 +1,11 @@
+from .annotation import *
+from .auto_labeling import *
+from .comment import *
+from .document import *
+from .health import *
+from .import_export import *
+from .label import *
+from .project import *
+from .role import *
+from .statistics import *
+from .user import *
