+++ app/api/views/project.py
@@ -1,0 +1,25 @@
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticated
+
+from ..models import Project
+from ..serializers import ProjectPolymorphicSerializer, ProjectSerializer
+from ..permissions import IsInProjectReadOnlyOrAdmin
+
+
+class ProjectList(generics.ListCreateAPIView):
+    serializer_class = ProjectPolymorphicSerializer
+    pagination_class = None
+    permission_classes = [IsAuthenticated & IsInProjectReadOnlyOrAdmin]
+
+    def get_queryset(self):
+        return self.request.user.projects
+
+    def perform_create(self, serializer):
+        serializer.save(users=[self.request.user])
+
+
+class ProjectDetail(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Project.objects.all()
+    serializer_class = ProjectSerializer
+    lookup_url_kwarg = 'project_id'
+    permission_classes = [IsAuthenticated & IsInProjectReadOnlyOrAdmin]
