+++ app/api/views/role.py
@@ -1,0 +1,35 @@
+from django.shortcuts import get_object_or_404
+from rest_framework import generics
+from rest_framework.permissions import IsAuthenticated
+
+from ..models import Role, Project, RoleMapping
+from ..permissions import IsProjectAdmin
+from ..serializers import RoleSerializer, RoleMappingSerializer
+
+
+class Roles(generics.ListCreateAPIView):
+    serializer_class = RoleSerializer
+    pagination_class = None
+    permission_classes = [IsAuthenticated & IsProjectAdmin]
+    queryset = Role.objects.all()
+
+
+class RoleMappingList(generics.ListCreateAPIView):
+    serializer_class = RoleMappingSerializer
+    pagination_class = None
+    permission_classes = [IsAuthenticated & IsProjectAdmin]
+
+    def get_queryset(self):
+        project = get_object_or_404(Project, pk=self.kwargs['project_id'])
+        return project.role_mappings
+
+    def perform_create(self, serializer):
+        project = get_object_or_404(Project, pk=self.kwargs['project_id'])
+        serializer.save(project=project)
+
+
+class RoleMappingDetail(generics.RetrieveUpdateDestroyAPIView):
+    queryset = RoleMapping.objects.all()
+    serializer_class = RoleMappingSerializer
+    lookup_url_kwarg = 'rolemapping_id'
+    permission_classes = [IsAuthenticated & IsProjectAdmin]
